Partial fix for dynamic classes
